update booking api
