+++ lib/booking_api/src/venue.rs
@@ -1,7 +1,7 @@
 use crate::service::VenueClient as Client;
 use async_trait::async_trait;
 use protobuf::venue::api::venue_api_client::VenueApiClient;
-use protobuf::venue::api::GetVenueRequest;
+use protobuf::venue::api::{GetVenueRequest, GetTablesRequest};
 use protobuf::venue::models::Venue;
 use tonic::transport::Channel;
 use tonic::Status;
@@ -24,5 +24,29 @@ impl Client for VenueClient {
             .get_venue(GetVenueRequest { id: venue_id })
             .await
             .map(|v| v.into_inner())
+    }
+
+    async fn get_tables_with_capacity(
+        &self,
+        venue_id: String,
+        capacity: u32,
+    ) -> Result<Vec<String>, Status> {
+        tracing::debug!(
+            "getting tables for venue {} with capacity {}",
+            &venue_id,
+            capacity
+        );
+
+        Ok(self
+            .client
+            .clone()
+            .get_tables(GetTablesRequest { venue_id })
+            .await?
+            .into_inner()
+            .tables
+            .iter()
+            .filter(|table| table.capacity >= capacity)
+            .map(|table| table.id.clone())
+            .collect())
     }
 }
